requirements.txt
python-decouple
pydrive
pyVmomi
tqdm
jinja2
pytz
csv-diff